refactor: move ChooseCollectionDropdown component
