+++ components/massmint/Massmint.vue
@@ -82,7 +82,7 @@ import { usePreferencesStore } from '@/stores/preferences'
 import UploadMediaZip from './uploadCompressedMedia/UploadCompressedMedia.vue'
 import UploadDescription from './uploadDescription/UploadDescription.vue'
 import OverviewTable from './OverviewTable.vue'
-import ChooseCollectionDropdown from './ChooseCollectionDropdown.vue'
+import ChooseCollectionDropdown from '@/components/common/ChooseCollectionDropdown.vue'
 import EditPanel from './EditPanel.vue'
 import { NFT, NFTToMint } from './types'
 import MissingInfoModal from './modals/MissingInfoModal.vue'
